backend/requirements.txt
flask
flask-cors
python-dotenv
google-genai
elevenlabs
# Audio
pyaudio
sounddevice
numpy
SpeechRecognition
# Vision
ultralytics
opencv-python
